outsource mockUpdateInfoServer to testhelper package
